+++ src/Symfony/Component/HttpFoundation/Session/Storage/MockArraySessionStorage.php
@@ -64,15 +64,30 @@ class MockArraySessionStorage implements SessionStorageInterface
     protected $bags;
 
     /**
+     * @var Boolean
+     */
+    protected $autoStart;
+
+    /**
+     * @var Boolean
+     */
+    protected $emulatePhp;
+
+    /**
      * Constructor.
      *
-     * @param string      $name    Session name
-     * @param MetadataBag $metaBag MetadataBag instance.
+     * @param string      $name       Session name.
+     * @param MetadataBag $metaBag    MetadataBag instance.
+     * @param Boolean     $autoStart  Autostart flag.
+     * @param Boolean     $emulatePhp Flag to allow access to session bags for inactive session
+     *                                when autostart is off.
      */
-    public function __construct($name = 'MOCKSESSID', MetadataBag $metaBag = null)
+    public function __construct($name = 'MOCKSESSID', MetadataBag $metaBag = null, $autoStart = true, $emulatePhp = true)
     {
         $this->name = $name;
         $this->setMetadataBag($metaBag);
+        $this->autoStart = $autoStart;
+        $this->emulatePhp = $emulatePhp;
     }
 
     /**
@@ -200,8 +215,10 @@ class MockArraySessionStorage implements SessionStorageInterface
             throw new \InvalidArgumentException(sprintf('The SessionBagInterface %s is not registered.', $name));
         }
 
-        if (!$this->started) {
+        if (!$this->started && $this->autoStart) {
             $this->start();
+        } elseif (!$this->started && !$this->autoStart && !$this->emulatePhp) {
+            throw new \RuntimeException('The session has not been started yet.');
         }
 
         return $this->bags[$name];