+++ src/Symfony/Component/HttpFoundation/Session/Storage/NativeSessionStorage.php
@@ -52,6 +52,16 @@ class NativeSessionStorage implements SessionStorageInterface
     protected $metadataBag;
 
     /**
+     * @var Boolean
+     */
+    protected $autoStart;
+
+    /**
+     * @var Boolean
+     */
+    protected $emulatePhp;
+
+    /**
      * Constructor.
      *
      * Depending on how you want the storage driver to behave you probably
@@ -91,11 +101,14 @@ class NativeSessionStorage implements SessionStorageInterface
      * upload_progress.min-freq, "1"
      * url_rewriter.tags, "a=href,area=href,frame=src,form=,fieldset="
      *
-     * @param array       $options Session configuration options.
-     * @param object      $handler SessionHandlerInterface.
-     * @param MetadataBag $metaBag MetadataBag.
+     * @param array       $options    Session configuration options.
+     * @param object      $handler    SessionHandlerInterface.
+     * @param MetadataBag $metaBag    MetadataBag.
+     * @param Boolean     $autoStart  Autostart flag.
+     * @param Boolean     $emulatePhp Flag to allow access to session bags for inactive session
+     *                                when autostart is off.
      */
-    public function __construct(array $options = array(), $handler = null, MetadataBag $metaBag = null)
+    public function __construct(array $options = array(), $handler = null, MetadataBag $metaBag = null, $autoStart = true, $emulatePhp = true)
     {
         ini_set('session.cache_limiter', ''); // disable by default because it's managed by HeaderBag (if used)
         ini_set('session.use_cookies', 1);
@@ -109,6 +122,8 @@ class NativeSessionStorage implements SessionStorageInterface
         $this->setMetadataBag($metaBag);
         $this->setOptions($options);
         $this->setSaveHandler($handler);
+        $this->autoStart = $autoStart;
+        $this->emulatePhp = $emulatePhp;
     }
 
     /**
@@ -258,8 +273,10 @@ class NativeSessionStorage implements SessionStorageInterface
 
         if ($this->saveHandler->isActive() && !$this->started) {
             $this->loadSession();
-        } elseif (!$this->started) {
+        } elseif (!$this->started && $this->autoStart) {
             $this->start();
+        } elseif (!$this->started && !$this->autoStart && !$this->emulatePhp) {
+            throw new \RuntimeException('The session has not been started yet.');
         }
 
         return $this->bags[$name];